Introduce -S option to make -t look only at start of flow
The -t <timewin> option only includes flows if both the flow's start
and end are inside timewin.

All of these:
http://sourceforge.net/p/nfdump/bugs/22/
http://sourceforge.net/p/nfdump/bugs/25/
and the email in nfdump-discuss with subject:
"How to use the nfdump -t <timewin> option for RRDTool integration?"

describe the same thing.  They would like the -t option to only
consider one end of the flow. See one of them for a more thorough
explanation.

This patch achieves that. It introduces an -S option, that only considers
whether the flow's start is inside timewin.

Instead of this pseudo-code:

    consider(master_record->first < twin_start)
    consider(master_record->last > twin_end)

does this pseudo-code:

    consider(master_record->first < twin_start)
    if (-S option)
        consider(master_record->first >=twin_end)
    else
        consider(master_record->last >= twin_end)

Beware, that in the getopt() line, all the options are listed in one
long line. That is heavily prone to merge conflicts. Also, I've chosen
the -S option, since it was available and a handy mnemonic for
"start".

The -S option was introduced for backwards-compatibility, so the -t
alone continues to work as it has always done.